start separating gameplay functions
