+++ Assets/Script/Model/BaseballModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9572b5eba0dcbf047b43d4cd4a04a537
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
