+++ Assets/Script/Model/GameModel.cs
@@ -2,8 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+//GamePlayModel, BaseballModel => 공통 변수, BattingModel, Pitching<odel 
 //Model
 public class GameModel
 {
-    
 }
